Updating comments sends the correct info - not yet tested with backend
The comment update form is now sending modified formData that contains what the CommentSchema wants (the comment content and its post id), along with the comment id via params.
The backend is currently giving a 500 error, but I believe this is because the update comments route is still in progress. This should be tested once the backend is ready, and may require changes depending on the result.

Work in progress for #51