PostView uses the new editable field to show and hide forms properly
If a user is signed in and owns the post, they can see the post edit and delete actions. If the user does not own the post, or is not signed in, they dont see those functions. Same logic with a user owning comments and comment update/delete. Also if a user is not signed, dont show add coment form.